@@ -13,11 +13,11 @@
   ],
   "repository": "GerHobbelt/markdown-it-anchor",
   "license": "UNLICENSE",
+  "type": "module",
   "source": "index.js",
   "main": "dist/markdownItAnchor.js",
-  "module": "dist/markdownItAnchor.mjs",
   "engines": {
-    "node": ">=10"
+    "node": ">=12.20.0"
   },
   "scripts": {
     "build": "make",
@@ -35,14 +35,15 @@
     "access": "public"
   },
   "devDependencies": {
-    "@gerhobbelt/markdown-it": "12.0.0-47",
+    "@gerhobbelt/markdown-it": "12.0.4-52",
     "@gerhobbelt/markdown-it-attrs": "3.0.3-20",
     "@gerhobbelt/markdown-it-testgen": "0.1.6-17",
-    "@gerhobbelt/prepend-header": "1.0.10-4",
-    "cross-env": "7.0.2",
-    "eslint": "7.10.0",
-    "microbundle": "0.12.4",
-    "mocha": "8.1.3",
+    "@gerhobbelt/prepend-header": "1.1.0-7",
+    "argparse": "2.0.1",
+    "cross-env": "7.0.3",
+    "eslint": "7.19.0",
+    "microbundle": "0.13.0",
+    "mocha": "8.2.1",
     "nyc": "15.1.0"
   },
   "reporter": [
