@@ -1,10 +1,18 @@
 
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+// see https://nodejs.org/docs/latest-v13.x/api/esm.html#esm_no_require_exports_module_exports_filename_dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
 function camelCase(str) {
   return str.replace(/-([\w])/g, (m, p1) => p1.toUpperCase());
 }
 
 const year = new Date().getFullYear();
-const pkg = require('../package.json');
+const pkg = JSON.parse(fs.readFileSync(path.normalize(path.join(__dirname, '../package.json')), 'utf8'));
 const version = pkg.version;
 const name = pkg.name.replace(/^.*?\//, '');
 const globalName = ((name) => {
@@ -30,7 +38,8 @@ const license = pkg.license;
 
 const text = `/*! ${name} ${version} https://github.com//GerHobbelt/${name} @license ${license} */\n\n`;
 const match = `/*! ${name} `; // skip the file where this match is true
-module.exports = {
+
+export default {
   text,
   match,
   version,
